update gameview.
If three players login, system automaticaly allocate seat for each
player and deliver cards, view can be updated when deliver cards, but
the performance is not good, the deliver cards process is not smooth,
need verify.
next step is to hook up the take card out into the system to run the
real play.