upload the cocos2d (html5 version)
We plan to replace the paperJS with cocos2d.

paperJS is desgined for a 2D sketch drawing app. It's not suitable for a
game develop.

cocos2d is very powerful on game develop aspect. I think we can get a
big benefit from it. Looking forward we can get further steps based on
the new framework.